notes: Remove course notes assignment from repo
For grading the professor is going to want the git log of the course
notes, in order to have a clean log I have separated the notes from the
rest of the assignments.

As these notes are not sensitive I will not worry about rebasing to
erase them from this repo entirely.